% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styling.R
\name{add_styles}
\alias{add_styles}
\title{Append new styles to CSS}
\usage{
add_styles(css, new_css = NULL)
}
\arguments{
\item{css}{The CSS to which new_styles or new_css will be added}
}
\value{
Updated \code{css} with new styles appended
}
\description{
The default new styles are internal data (new_styles in R/sysdata.rda), but
\code{new_css} can be specified for users to append their own CSS.
}
\seealso{
if any see alsos
}

