% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{add_yaml}
\alias{add_yaml}
\title{Add rmarkdown YAML front-matter to markdown doc}
\usage{
add_yaml(doc, theme = "default", styles = list(), shiny = FALSE,
  yml = NULL)
}
\arguments{
\item{doc}{The markdown document to append to YAML}

\item{theme}{A Bootswatch theme (see \url{http://bootswatch.com}) ["default"]}

\item{styles}{List of paths to one or more CSS style sheets}

\item{shiny}{Whether to use the Shiny runtime [FALSE]}

\item{yml}{Optional user-supplied YAML frontmatter [NULL]}
}
\value{
An Rmarkdown version of \code{doc} with a YAML header
}
\description{
Add rmarkdown YAML front-matter to markdown doc
}
\details{
Rmarkdown documents take a YAML header to describe output options.
This function adds the basic header:

\preformatted{---
output:
  html_document:
    toc: true
    toc_float: true
    css: [path/to/css1, path/to/css2]
---}

Users may supply any CSS they wish; an example is provided with
\link{use_example_css} (and write with \link{write_example_css}. Users may
supply their own YAML header, but must #' correctly specify the (absolute)
path to the CSS file.
}

