% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_cleanup.R
\name{clean_brackets}
\alias{clean_brackets}
\title{Un-escape square brackets}
\usage{
clean_brackets(doc)
}
\arguments{
\item{doc}{The markdown text to be cleaned}
}
\value{
A clean version of doc
}
\description{
Pandoc escapes square brackets in the HTML:md conversion, which makes the
Rmarkdown invalid. This function un-escapes the brackets.
}
\examples{
\dontrun{
doc <- "This is the issue \\\\[here\\\\]"
clean_brackets(doc)
}
}

