% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_mods.R
\name{clean_hrefs}
\alias{clean_hrefs}
\title{Remove extraneous strings from URLs}
\usage{
clean_hrefs(html)
}
\arguments{
\item{html}{The HTML to be searched and sub'd}
}
\value{
An updated form of \code{html}
}
\description{
Google adds a bunch of its own modifiers to URLs in GDocs, and these modifiers
will break the links. This removes the modifiers.
}

