% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_mods.R
\name{convert_pelements}
\alias{convert_pelements}
\title{Insert <div>s in front of <p>s}
\usage{
convert_pelements(html)
}
\arguments{
\item{html}{The HTML whose <p>-elements will be converted}
}
\value{
An updated html page with <p><div>...</div></p>
}
\description{
\code{pandoc} converts HTML to Markdown in part by inserting newlines where
there are <p> (</p>) elements in HTML. When those <p> elements are refined
by a class specification then formatting is lost.
}
\details{
To keep the class details, this function inserts a <div> behind each <p>.
As a result, HTML such as:

\code{<p class="newstyle">This is a p-element.</p>}

becomes

\code{<p><div class="newstyle">This is a p-element</div></p>},

which then becomes

\code{\\n<div class="newstyle">This is a p-element</div>\\n}

after conversion to markdown and retains the information in class "newstyle".
}
\examples{
html <- <p class="newstyle">This is a p-element.</p>
convert_pelements(html)
}

