% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate.R
\name{decapitate}
\alias{decapitate}
\title{Separate the <head> from the <body> of an html file}
\usage{
decapitate(doc)
}
\arguments{
\item{doc}{An object representation of an html file}
}
\value{
A list with the <head> ($head) and <body> ($body) of \code{doc}
}
\description{
This function works line-wise to separate the <head> from the <body>,
so using, e.g., \code{readLines(infile)} as the input is suitable. Checks
if \code{length(doc) == 1} and splits on newlines to create the vector.
}
\examples{
\dontrun{
decapitate(html_file)
}
}

