% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_mods.R
\name{fix_html_escapes}
\alias{fix_html_escapes}
\title{Fix HTML escape codes for equals and space in URLs}
\usage{
fix_html_escapes(txt)
}
\arguments{
\item{html}{The HTML to be searched and sub'd}
}
\value{
An updated form of \code{html}
}
\description{
Fix HTML escape codes for equals and space in URLs
}

