% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_mods.R
\name{fix_title_span}
\alias{fix_title_span}
\title{Upate the GDoc title for proper formatting}
\usage{
fix_title_span(html)
}
\arguments{
\item{html}{The HTML to be searched and sub'd}
}
\value{
An updated form of \code{html}
}
\description{
Upate the GDoc title for proper formatting
}

