% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styling.R
\name{get_css}
\alias{get_css}
\title{Extract the <style> from HTML text}
\usage{
get_css(txt)
}
\arguments{
\item{txt}{An HTML <head> character vector}
}
\value{
CSS character string
}
\description{
Extract the <style> from HTML text
}
\seealso{
\link{update_css}
}

