% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html2md.R
\name{html2md}
\alias{html2md}
\title{Use \code{pandoc} to convert HTML to markdown}
\usage{
html2md(html, mdf)
}
\arguments{
\item{html}{The HTML to be searched and sub'd}
}
\value{
An updated form of \code{html}
}
\description{
Use \code{pandoc} to convert HTML to markdown
}

