% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styling.R
\name{remove_at_import}
\alias{remove_at_import}
\title{Remove google's \@import line from <style>}
\usage{
remove_at_import(css)
}
\arguments{
\item{css}{The CSS from which the import is stripped}
}
\value{
The clean CSS
}
\description{
Remove google's \@import line from <style>
}

