% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_mods.R
\name{remove_empties}
\alias{remove_empties}
\title{Remove empty heading elements}
\usage{
remove_empties(txt)
}
\arguments{
\item{html}{The HTML to be searched and sub'd}
}
\value{
An updated form of \code{html}
}
\description{
Remove empty heading elements
}

