% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_mods.R
\name{replace_empty_headings}
\alias{replace_empty_headings}
\title{Remove empty heading elements}
\usage{
replace_empty_headings(txt)
}
\arguments{
\item{html}{The HTML to be searched and sub'd}
}
\value{
An updated form of \code{html}
}
\description{
Remove empty heading elements
}

