% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_mods.R
\name{replace_gdoc_statement}
\alias{replace_gdoc_statement}
\title{Replace a specific string to reflect medium change}
\usage{
replace_gdoc_statement(html)
}
\arguments{
\item{html}{The HTML to be searched and sub'd}
}
\value{
An updated form of \code{html}
}
\description{
The string we are using is "This Google Doc", which is replaced with "This
page was written with Google Docs and converted using the `gdoc2rmd` R
package. It".
}

