% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styling.R
\name{update_css}
\alias{update_css}
\title{Update the CSS from GDoc HTML}
\usage{
update_css(txt, no_import = TRUE, wrap_css = TRUE, add_style = TRUE)
}
\arguments{
\item{txt}{An HTML <head> character vector}

\item{no_import}{Remove the google \@import statement from CSS? (TRUE)}

\item{wrap_css}{Wrap CSS for pretty-printing? (TRUE)}

\item{add_style}{Add new styles with \link{add_styles}? (TRUE)}
}
\value{
CSS character string
}
\description{
A wrapper to: extract the CSS from a Google Docs HTML export; remove the
\@import GDocs contain; wrap the extracted (inline) CSS; and add style
revisions
}
\seealso{
\link{update_css}
}

