% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_mods.R
\name{update_gdoc_statement}
\alias{update_gdoc_statement}
\title{Very specific string sub for our demo}
\usage{
update_gdoc_statement(html)
}
\arguments{
\item{html}{The HTML to be searched and sub'd}
}
\value{
An updated form of \code{html}
}
\description{
Very specific string sub for our demo
}

