% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styling.R
\name{use_example_css}
\alias{use_example_css}
\title{Use 'new_styles.css' from system.file}
\usage{
use_example_css()
}
\value{
Path to the system.file new_styles.css
}
\description{
The internal data 'new_styles' includes formatting for the major HTML
attributes, e.g., p, h1-6, li, ol. This data may included in, e.g.,
\link{add_default_yaml}.
}
\seealso{
\link{write_example_css}
}

