% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styling.R
\name{wrap_css}
\alias{wrap_css}
\title{Wrap CSS for pretty-printing}
\usage{
wrap_css(css)
}
\arguments{
\item{css}{The (inlined) CSS to be wrapped}
}
\value{
The wrapped CSS
}
\description{
Wrap CSS for pretty-printing
}

