% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml.R
\name{add_yaml}
\alias{add_yaml}
\title{Add rmarkdown YAML front-matter to markdown doc}
\usage{
add_yaml(doc, css = NULL, yml = NULL)
}
\arguments{
\item{doc}{The markdown document to append to YAML}

\item{css}{Path to CSS file to be included during knitting}

\item{yml}{Optional user-supplied YAML frontmatter (NULL)}
}
\value{
A base rmarkdown version of markdown \code{doc}
}
\description{
Add rmarkdown YAML front-matter to markdown doc
}
\details{
Rmarkdown documents require a YAML header to describe output types
and options. This function adds the basic header:

\preformatted{---
output:
  html_document:
    css: <path/to/css>
    toc: true
    toc_float: true
---}

Users may supply their own YAML header, but must correctly specify the
(absolute) path to the CSS file.
}

